backend/requirements.txt
Django==4.2.25
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
PyJWT==2.8.0
python-decouple==3.8
dj-database-url==2.1.0
whitenoise==6.5.0